pyproject.toml
[tool.poetry]
name = "signifi1-0"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "signifi1"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
